Refactored `common` files to flow with split
