+++ gender_analysis/gender.py
@@ -1,5 +1,6 @@
 from gender_analysis.pronouns import PronounSeries
 
+
 class Gender:
     """
     Defines a gender that will be operated on in analysis functions
